requirements.txt
# Core Dependencies
fastapi
uvicorn
streamlit
pydantic
python-multipart
requests

# AI/ML Libraries
langchain
langgraph
langchain-core
langchain-google-genai
langchain-community

# MCP Dependencies
mcp
httpx
langchain-mcp-adapters

# Testing
pytest
pytest-asyncio
pytest-mock

# Utilities
python-dotenv
aiohttp 